Vendredi 18/06 matin before fussioner the  tables address and person

+++ src/main/java/com/simplon/safety/alertsnet/dao/AddressDao.java
@@ -10,7 +10,7 @@ public interface AddressDao {
 	
 	int initAddressTable() throws IOException;
 	
-	int insertAddress(Address address) throws IOException;
+	long insertAddress(Address address) throws IOException;
 	
 	List<Address> listAllAddress() throws IOException;
